+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `localOptimizationStatistics.download(basePath: string, includedFilterIds, excludedFilterIds)`:
+  Allow both `includedFilterIds` and `excludedFilterIds` to be used at once.
 - `localOptimizationStatistics.download(basePath)`: Clear the `basePath` directory before downloading statistics.
 
 ### Deprecated